requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-ipware==3.0.2
idna==2.10
install==1.3.4
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
psycopg2==2.8.6
ptvsd==4.3.2
pycodestyle==2.7.0
pylint==2.8.2
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
wrapt==1.12.1
